DTSCCCI-6067: move pages/actions under common folder
