+++ playwright-e2e/actions/ui/exui/defendant-solicitor/defendant-actions-factory.ts
@@ -17,8 +17,8 @@ import UploadMediationDocumentsActions from '../claimant-solicitor/upload-mediat
 import UploadMediationDocumentsPageFactory from '../../../../pages/exui/claimant-defendant-solicitor/upload-mediation-documents/upload-mediation-documents-page-factory';
 import QueryManagementActions from '../common/query-management/query-management-actions.ts';
 import QueryManagementPageFactory from '../../../../pages/exui/common/query-management/query-management-page-factory.ts';
-import RefundActions from '../refunds/refund-actions';
-import RefundPageFactory from '../../../../pages/exui/refunds/refund-page-factory';
+import RefundActions from '../common/refunds/refund-actions';
+import RefundPageFactory from '../../../../pages/exui/common/refunds/refund-page-factory';
 
 export default class DefendantActionsFactory extends BasePageActionsFactory {
   get defendantResponseActions() {
