+++ playwright-e2e/pages/exui/common/refunds/refund-confirm/refund-confirm-rejected-page.ts
@@ -1,5 +1,5 @@
-import BasePage from '../../../../base/base-page';
-import { AllMethodsStep } from '../../../../decorators/test-steps';
+import BasePage from '../../../../../base/base-page';
+import { AllMethodsStep } from '../../../../../decorators/test-steps';
 import { headings } from './refund-confirm-content';
 
 @AllMethodsStep()
