+++ playwright-e2e/pages/exui/common/refunds/service-request/service-request-content.ts
