+++ playwright-e2e/tests/ui-tests/refunds/1v1_request_for_refunds.test.ts
@@ -1,0 +1,51 @@
+import { test } from '../../../playwright-fixtures/index';
+
+test.describe(
+  '1v1 create fast claim - request and processing of refunds',
+  { tag: '@civil-ccd-nightly' },
+  () => {
+    test('request and approve a returned refund', async ({
+      ClaimantSolicitorApiSteps,
+      CaseRoleAssignmentApiSteps,
+      ClaimantSolicitorSteps,
+      DefendantSolicitor1Steps,
+    }) => {
+      await ClaimantSolicitorApiSteps.CreateClaimFast1v1();
+      await ClaimantSolicitorApiSteps.MakePaymentForClaimIssue();
+      await ClaimantSolicitorApiSteps.NotifyClaim();
+      await CaseRoleAssignmentApiSteps.AssignCaseRoleToDS1();
+      await ClaimantSolicitorApiSteps.NotifyClaimDetails();
+      await ClaimantSolicitorApiSteps.MakePaymentForRefund();
+      await ClaimantSolicitorSteps.Login();
+      await ClaimantSolicitorSteps.RequestRefund();
+      await DefendantSolicitor1Steps.Login();
+      await DefendantSolicitor1Steps.ReturnRefundToCaseworker();
+      await ClaimantSolicitorSteps.Login();
+      await ClaimantSolicitorSteps.AmendReturnedRefund();
+      await DefendantSolicitor1Steps.Login();
+      await DefendantSolicitor1Steps.ApproveRefund();
+    });
+
+    test('request and reject a returned refund', async ({
+      ClaimantSolicitorApiSteps,
+      CaseRoleAssignmentApiSteps,
+      ClaimantSolicitorSteps,
+      DefendantSolicitor1Steps,
+    }) => {
+      await ClaimantSolicitorApiSteps.CreateClaimFast1v1();
+      await ClaimantSolicitorApiSteps.MakePaymentForClaimIssue();
+      await ClaimantSolicitorApiSteps.NotifyClaim();
+      await CaseRoleAssignmentApiSteps.AssignCaseRoleToDS1();
+      await ClaimantSolicitorApiSteps.NotifyClaimDetails();
+      await ClaimantSolicitorApiSteps.MakePaymentForRefund();
+      await ClaimantSolicitorSteps.Login();
+      await ClaimantSolicitorSteps.RequestRefund();
+      await DefendantSolicitor1Steps.Login();
+      await DefendantSolicitor1Steps.ReturnRefundToCaseworker();
+      await ClaimantSolicitorSteps.Login();
+      await ClaimantSolicitorSteps.AmendReturnedRefund();
+      await DefendantSolicitor1Steps.Login();
+      await DefendantSolicitor1Steps.RejectRefund();
+    });
+  },
+);
